backend/pyproject.toml
[project]
name = "suit-crm"
version = "0.1.0"
description = "A CRM system for managing suit orders"
authors = [
    { name = "Ryan", email = "your.email@example.com" }
]
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "sqlalchemy==2.0.23",
    "pydantic==2.4.2",
    "python-dotenv==1.0.0",
    "pyodbc==4.0.39",
    "Flask==2.3.3",
    "Flask-CORS==4.0.0",
    "gunicorn==21.2.0"
]
requires-python = ">=3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 